backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx==0.24.1